claude-skills/py-exchangelib/requirements.txt
exchangelib>=5.0.0
python-dotenv>=1.0.0
msal>=1.30.0
